Move helper methods into util.rb
