+++ lib/rulers.rb
@@ -1,6 +1,5 @@
-require 'rulers/array'
-require 'rulers/object'
 require 'rulers/routing'
+require 'rulers/util'
 require 'rulers/version'
 
 module Rulers
