--- lib/rulers/array.rb
@@ -1,2 +1,0 @@
-class Array
-end
