+++ lib/rulers/util.rb
